Ongoing Test Fixes - Test Setup Refactor
