+++ tests/block_test_util.go
@@ -115,8 +115,6 @@ func (t *BlockTest) Run() error {
 		return err
 	}
 	defer chain.Stop()
-	// @SHYFT NOTE: Clear pg database
-	// core.TruncateTables()
 	validBlocks, err := t.insertBlocks(chain)
 	if err != nil {
 		return err
